+++ packages/notebook/src/browser/view-model/notebook-cell-model.ts
@@ -100,6 +100,9 @@ export class NotebookCellModel implements NotebookCell, Disposable {
     protected readonly onDidRequestCellEditChangeEmitter = new Emitter<boolean>();
     readonly onDidRequestCellEditChange = this.onDidRequestCellEditChangeEmitter.event;
 
+    protected readonly onWillFocusCellEditorEmitter = new Emitter<void>();
+    readonly onWillFocusCellEditor = this.onWillFocusCellEditorEmitter.event;
+
     @inject(NotebookCellModelProps)
     protected readonly props: NotebookCellModelProps;
 
@@ -114,6 +117,11 @@ export class NotebookCellModel implements NotebookCell, Disposable {
 
     get metadata(): NotebookCellMetadata {
         return this._metadata;
+    }
+
+    set metadata(newMetadata: NotebookCellMetadata) {
+        this._metadata = newMetadata;
+        this.onDidChangeMetadataEmitter.fire();
     }
 
     protected _metadata: NotebookCellMetadata;
@@ -212,6 +220,11 @@ export class NotebookCellModel implements NotebookCell, Disposable {
     requestStopEdit(): void {
         this._editing = false;
         this.onDidRequestCellEditChangeEmitter.fire(false);
+    }
+
+    requestFocusEditor(): void {
+        this.requestEdit();
+        this.onWillFocusCellEditorEmitter.fire();
     }
 
     spliceNotebookCellOutputs(splice: NotebookCellOutputsSplice): void {