+++ packages/terminal/src/browser/terminal-frontend-module.ts
@@ -41,7 +41,7 @@ import { TerminalThemeService } from './terminal-theme-service';
 import { QuickAccessContribution } from '@theia/core/lib/browser/quick-input/quick-access';
 import { createXtermLinkFactory, TerminalLinkProvider, TerminalLinkProviderContribution, XtermLinkFactory } from './terminal-link-provider';
 import { UrlLinkProvider } from './terminal-url-link-provider';
-import { FileDiffPostLinkProvider, FileDiffPreLinkProvider, FileLinkProvider } from './terminal-file-link-provider';
+import { FileDiffPostLinkProvider, FileDiffPreLinkProvider, FileLinkProvider, LocalFileLinkProvider } from './terminal-file-link-provider';
 import {
     ContributedTerminalProfileStore, DefaultProfileStore, DefaultTerminalProfileService,
     TerminalProfileService, TerminalProfileStore, UserTerminalProfileStore
@@ -123,6 +123,8 @@ export default new ContainerModule(bind => {
     bind(TerminalLinkProvider).toService(FileDiffPreLinkProvider);
     bind(FileDiffPostLinkProvider).toSelf().inSingletonScope();
     bind(TerminalLinkProvider).toService(FileDiffPostLinkProvider);
+    bind(LocalFileLinkProvider).toSelf().inSingletonScope();
+    bind(TerminalLinkProvider).toService(LocalFileLinkProvider);
 
     bind(ContributedTerminalProfileStore).to(DefaultProfileStore).inSingletonScope();
     bind(UserTerminalProfileStore).to(DefaultProfileStore).inSingletonScope();